requirements.txt
camelot-py
PyPDF2
Flask
flask-cors
google-cloud-aiplatform
supabase